[issue-406] rename json parser package to jsonlikedict
