+++ src/parser/jsonlikedict/actor_parser.py
@@ -13,7 +13,7 @@ from typing import Pattern, Match, Optional
 
 from src.model.actor import Actor, ActorType
 from src.parser.error import SPDXParsingError
-from src.parser.json.dict_parsing_functions import construct_or_raise_parsing_error
+from src.parser.jsonlikedict.dict_parsing_functions import construct_or_raise_parsing_error
 
 
 class ActorParser:
