+++ src/parser/jsonlikedict/creation_info_parser.py
@@ -17,9 +17,9 @@ from src.model.document import CreationInfo
 from src.model.external_document_ref import ExternalDocumentRef
 from src.model.version import Version
 from src.parser.error import SPDXParsingError
-from src.parser.json.actor_parser import ActorParser
-from src.parser.json.checksum_parser import ChecksumParser
-from src.parser.json.dict_parsing_functions import append_parsed_field_or_log_error, \
+from src.parser.jsonlikedict.actor_parser import ActorParser
+from src.parser.jsonlikedict.checksum_parser import ChecksumParser
+from src.parser.jsonlikedict.dict_parsing_functions import append_parsed_field_or_log_error, \
     raise_parsing_error_if_logger_has_messages, construct_or_raise_parsing_error, parse_field_or_log_error, \
     parse_field_or_no_assertion
 from src.datetime_conversions import datetime_from_str
