+++ src/parser/jsonlikedict/license_expression_parser.py
@@ -12,7 +12,7 @@ from typing import Union, List
 
 from src.model.license_expression import LicenseExpression
 from src.parser.error import SPDXParsingError
-from src.parser.json.dict_parsing_functions import construct_or_raise_parsing_error, append_parsed_field_or_log_error, \
+from src.parser.jsonlikedict.dict_parsing_functions import construct_or_raise_parsing_error, append_parsed_field_or_log_error, \
     raise_parsing_error_if_logger_has_messages
 from src.parser.logger import Logger
