+++ tests/parser/test_checksum_parser.py
@@ -14,7 +14,7 @@ import pytest
 
 from src.model.checksum import ChecksumAlgorithm
 from src.parser.error import SPDXParsingError
-from src.parser.json.checksum_parser import ChecksumParser
+from src.parser.jsonlikedict.checksum_parser import ChecksumParser
 
 
 def test_parse_checksum():
