+++ tests/parser/test_extracted_licensing_info_parser.py
@@ -13,7 +13,7 @@ from unittest import TestCase
 import pytest
 
 from src.parser.error import SPDXParsingError
-from src.parser.json.extracted_licensing_info_parser import ExtractedLicensingInfoParser
+from src.parser.jsonlikedict.extracted_licensing_info_parser import ExtractedLicensingInfoParser
 
 
 def test_parse_extracted_licensing_info():
