+++ tests/parser/test_file_parser.py
@@ -16,8 +16,8 @@ from src.model.checksum import Checksum, ChecksumAlgorithm
 from src.model.file import FileType
 from src.model.license_expression import LicenseExpression
 from src.parser.error import SPDXParsingError
-from src.parser.json.dict_parsing_functions import parse_list_of_elements
-from src.parser.json.file_parser import FileParser
+from src.parser.jsonlikedict.dict_parsing_functions import parse_list_of_elements
+from src.parser.jsonlikedict.file_parser import FileParser
 
 
 def test_parse_file():
